requirements.txt
pandas
numpy
requests
python-dotenv
pyarrow

scikit-learn
joblib

streamlit
plotly
prophet

shap==0.45.0
numba==0.58.1
llvmlite==0.41.1
lime

pytest
black
